backend/requirements.txt
Django>=5.0,<5.1
djangorestframework
django-cors-headers
psycopg2-binary
Pillow
python-dotenv
django-filter
